Allow changing highlighter versions
